--- projects/rocr-runtime/libhsakmt/include/impl/wddm/ntstatus_translate.h
@@ -1,84 +1,0 @@
-////////////////////////////////////////////////////////////////////////////////
-//
-// The University of Illinois/NCSA
-// Open Source License (NCSA)
-//
-// Copyright (c) 2020-2026, Advanced Micro Devices, Inc. All rights reserved.
-//
-// Developed by:
-//
-//                 AMD Research and AMD HSA Software Development
-//
-//                 Advanced Micro Devices, Inc.
-//
-//                 www.amd.com
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to
-// deal with the Software without restriction, including without limitation
-// the rights to use, copy, modify, merge, publish, distribute, sublicense,
-// and/or sell copies of the Software, and to permit persons to whom the
-// Software is furnished to do so, subject to the following conditions:
-//
-//  - Redistributions of source code must retain the above copyright notice,
-//    this list of conditions and the following disclaimers.
-//  - Redistributions in binary form must reproduce the above copyright
-//    notice, this list of conditions and the following disclaimers in
-//    the documentation and/or other materials provided with the distribution.
-//  - Neither the names of Advanced Micro Devices, Inc,
-//    nor the names of its contributors may be used to endorse or promote
-//    products derived from this Software without specific prior written
-//    permission.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
-// THE CONTRIBUTORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR
-// OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
-// ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
-// DEALINGS WITH THE SOFTWARE.
-//
-////////////////////////////////////////////////////////////////////////////////
-
-#pragma once
-
-#if defined(__linux__)
-#include "impl/wddm/wddm_types.h"
-#else
-#define WIN32_NO_STATUS
-#include <Windows.h>
-#undef WIN32_NO_STATUS
-#endif
-
-#include "ntstatus.h"
-#include "impl/wddm/status.h"
-
-namespace wsl {
-namespace thunk {
-
-inline ErrorCode TranslateNtStatus(NTSTATUS status) {
-  switch (status) {
-  case STATUS_SUCCESS:
-    return ErrorCode::Success;
-  case STATUS_PENDING:
-    return ErrorCode::NotReady;
-  case STATUS_NO_MEMORY:
-     return ErrorCode::OutOfMemory;
-  case STATUS_DEVICE_REMOVED:
-    return ErrorCode::DeviceLost;
-  case STATUS_GRAPHICS_NO_VIDEO_MEMORY:
-    return ErrorCode::OutOfGpuMemory;
-  case STATUS_TIMEOUT:
-    return ErrorCode::Timeout;
-  case STATUS_INVALID_PARAMETER:
-    return ErrorCode::InvalidateParams;
-  case STATUS_INVALID_HANDLE:
-    return ErrorCode::InvalidHandle;
-  default:
-    break;
-  }
-  return ErrorCode::Unknown;
-}
-
-} // namespace thunk
-} // namespace wsl
