+++ projects/rocr-runtime/libhsakmt/src/dxg/wddm/device.cpp
@@ -52,7 +52,6 @@
 #include <sys/stat.h>
 #include <fcntl.h>
 #include "impl/wddm/status.h"
-#include "impl/wddm/ntstatus_translate.h"
 #include "impl/wddm/types.h"
 #include "impl/wddm/device.h"
 #include "impl/wddm/queue.h"
@@ -165,12 +164,7 @@ bool WDDMDevice::QuerySegmentInfo()
 
 bool WDDMDevice::FindSegmentId(SegmentKind segment_kind, uint32_t* segment_id)
 {
-  if (segment_id == nullptr) {
-    return false;
-  }
-
   for (const auto& seg_info : segment_infos_) {
-
     if (seg_info.kind == segment_kind) {
       *segment_id = seg_info.segment_id;
       return true;
@@ -183,7 +177,7 @@ bool WDDMDevice::FindSegmentId(SegmentKind segment_kind, uint32_t* segment_id)
 /*Local heap(dedicated GPU memory) includes visible heap and invisible heap.
  *Non local heap refers to shared GPU memory and it is system memory.
  */
-ErrorCode WDDMDevice::VramAvail(uint64_t* available_bytes) {
+hsa_status_t WDDMDevice::VramAvail(uint64_t* available_bytes) {
   D3DKMT_QUERYSTATISTICS stats;
   NTSTATUS ret;
   uint64_t usedVis = 0;
@@ -191,36 +185,25 @@ ErrorCode WDDMDevice::VramAvail(uint64_t* available_bytes) {
   uint64_t usedNonLocal = 0;
   uint32_t segmentId = 0;
 
-  if (!available_bytes) {
-    return ErrorCode::InvalidateParams;
-  }
-
   *available_bytes = 0;
 
   // wait fence complete
   uint64_t value = page_fence_value_.load();
-  if (!CpuWait(&page_syncobj_, &value, 1, false)) {
-    pr_err("Failed to wait for paging fence\n");
-    return ErrorCode::Unknown;
-  }
+  if (!CpuWait(&page_syncobj_, &value, 1, false))
+    return HSA_STATUS_ERROR;
 
   if (IsDgpu()) {
     // local cpu-visible memory
-    if (!FindSegmentId(SegmentKind::kLocalMemory, &segmentId)) {
-      pr_err("Failed to find local memory segment\n");
-      return ErrorCode::Unknown;
-    }
+    if (!FindSegmentId(SegmentKind::kLocalMemory, &segmentId))
+      return HSA_STATUS_ERROR;
 
     memset(&stats, 0, sizeof(D3DKMT_QUERYSTATISTICS));
     stats.Type = D3DKMT_QUERYSTATISTICS_SEGMENT;
     stats.AdapterLuid = adapter_luid_;
     stats.QuerySegment.SegmentId = segmentId;
     ret = DXCORE_CALL(D3DKMTQueryStatistics(&stats));
-    if (ret != STATUS_SUCCESS) {
-      pr_err("Failed to query visible local memory segment %u info: %x\n", segmentId, ret);
-      return TranslateNtStatus(ret);
-    }
-    usedVis = stats.QueryResult.SegmentInformation.BytesResident;
+    if (ret == 0)
+      usedVis = stats.QueryResult.SegmentInformation.BytesResident;
 
     // local invisible memory
     if (device_info_.local_invisible_heap_size) {
@@ -237,8 +220,7 @@ ErrorCode WDDMDevice::VramAvail(uint64_t* available_bytes) {
       }
 
       if (!foundInvisible) {
-        pr_err("Failed to find invisible memory segment\n");
-        return ErrorCode::Unknown;
+        return HSA_STATUS_ERROR;
       }
       memset(&stats, 0, sizeof(D3DKMT_QUERYSTATISTICS));
       stats.Type = D3DKMT_QUERYSTATISTICS_SEGMENT;
@@ -246,37 +228,28 @@ ErrorCode WDDMDevice::VramAvail(uint64_t* available_bytes) {
       stats.QuerySegment.SegmentId = invisibleSegmentId;
 
       ret = DXCORE_CALL(D3DKMTQueryStatistics(&stats));
-      if (ret != STATUS_SUCCESS) {
-        pr_err("Failed to query invisible local memory segment %u info: %x\n",
-               invisibleSegmentId, ret);
-        return TranslateNtStatus(ret);
-      }
-      usedInv = stats.QueryResult.SegmentInformation.BytesResident;
+      if (ret == 0)
+        usedInv = stats.QueryResult.SegmentInformation.BytesResident;
     }
 
     *available_bytes = LocalHeapSize() - usedVis - usedInv;
   } else {
     // APU - NonLocal memory
-    if (!FindSegmentId(SegmentKind::kSystemMemory, &segmentId)) {
-      pr_err("Failed to find system memory segment\n");
-      return ErrorCode::Unknown;
-    }
+    if (!FindSegmentId(SegmentKind::kSystemMemory, &segmentId))
+      return HSA_STATUS_ERROR;
 
     memset(&stats, 0, sizeof(D3DKMT_QUERYSTATISTICS));
     stats.Type = D3DKMT_QUERYSTATISTICS_SEGMENT;
     stats.AdapterLuid = adapter_luid_;
     stats.QuerySegment.SegmentId = segmentId;
     ret = DXCORE_CALL(D3DKMTQueryStatistics(&stats));
-    if (ret != STATUS_SUCCESS) {
-      pr_err("Failed to query system memory segment %u info: %x\n", segmentId, ret);
-      return TranslateNtStatus(ret);
-    }
-    usedNonLocal = stats.QueryResult.SegmentInformation.BytesResident;
+    if (ret == 0)
+      usedNonLocal = stats.QueryResult.SegmentInformation.BytesResident;
 
     *available_bytes = NonLocalHeapSize() - usedNonLocal;
   }
 
-  return ErrorCode::Success;
+  return HSA_STATUS_SUCCESS;
 }
 
 bool WDDMDevice::CreateDevice(void) {
